+++ Cargo.toml
@@ -8,9 +8,9 @@ edition = "2018"
 byteorder = { version = "1.4.3", default-features = false }
 scale-info = { version = "0.10.0", default-features = false }
 serde = { version = "1.0.126", default-features = false }
-serde_json = "1.0.66"
 
 [dev-dependencies]
 parity-scale-codec = { version = "2.2.0", features = ["derive"] }
 scale-info = { version = "0.10.0", features = ["derive"] }
 serde = { version = "1.0.126", features = ["derive"] }
+serde_json = "1.0.66"
